Global syntax rewrite. I fuck myself for being picky about codestyle

+++ build.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
-rm -rf build
+if [ -d build ]; then
+  rm -rf build
+fi
+
 cmake -S . -B build
 cmake --build build
